+++ networking/eth2/src/integration-test/java/tech/pegasys/teku/networking/eth2/AbstractRpcMethodIntegrationTest.java
@@ -16,16 +16,22 @@ package tech.pegasys.teku.networking.eth2;
 import static org.assertj.core.util.Preconditions.checkState;
 
 import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.params.provider.Arguments;
+import tech.pegasys.teku.infrastructure.async.Waiter;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.eth2.peers.Eth2Peer;
 import tech.pegasys.teku.networking.eth2.rpc.core.encodings.RpcEncoding;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.TestSpecFactory;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
+import tech.pegasys.teku.spec.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair.BeaconBlockBodyAltair;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.bellatrix.BeaconBlockBodyBellatrix;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.capella.BeaconBlockBodyCapella;
@@ -240,6 +246,31 @@ public abstract class AbstractRpcMethodIntegrationTest {
 
   protected static Stream<Arguments> generateSpec() {
     return Arrays.stream(SpecMilestone.values()).map(Arguments::of);
+  }
+
+  protected List<BlobSidecar> retrieveCanonicalBlobSidecarsFromPeerStorage(
+      final Stream<UInt64> slots) {
+
+    return slots
+        .map(
+            slot ->
+                peerStorage
+                    .recentChainData()
+                    .getBlockRootBySlot(slot)
+                    .map(root -> new SlotAndBlockRoot(slot, root)))
+        .filter(Optional::isPresent)
+        .map(Optional::get)
+        .map(this::safeRetrieveBlobSidecars)
+        .flatMap(Collection::stream)
+        .collect(Collectors.toUnmodifiableList());
+  }
+
+  private List<BlobSidecar> safeRetrieveBlobSidecars(final SlotAndBlockRoot slotAndBlockRoot) {
+    try {
+      return Waiter.waitFor(peerStorage.recentChainData().retrieveBlobSidecars(slotAndBlockRoot));
+    } catch (Exception e) {
+      throw new RuntimeException(e);
+    }
   }
 
   protected static Class<?> milestoneToBeaconBlockBodyClass(final SpecMilestone milestone) {
